Add files via upload
Just put the name of your scripts into this file. This will run all of the programs made assuming the names of each script are in the file. Also, if data is needed from other scripts, just put:
Group_Project;
at the start of your script to import the variables from all of the scripts.